Add files via upload
DATA VISUALISATION FOR WAGGLE PET TRACKER DEVICE

Waggle is a startup that makes smart devices for pets. This project aims to tell a story of how the Lapcat data compares to findings from the dog collar Lapdog devices. 


REPORT FEATURES
- Report includes a custom Power BI theme that incorporates the provided branding colors
- Report includes slicers on each page that users can use to filter the report. The slicer types are; A drop-down slicer, a slider slicer, a hierarchy slicer, a slicer with a search box, and a slicer with “select all” enabled.
- Report includes a bookmark that swaps one visual out with a different visual
- Report includes a bookmark that resets all filters on a page to their default state.
- Report includes a button for triggering the filter reset bookmark. It alerts the user by changing size or color when users hover.
- The buttons for triggering the bookmarks indicate which bookmark is active or inactive.
- Report includes buttons that can be used to navigate the report.
- All buttons responds to interaction when a user hovers over them by changing color and slightly changing size or thickness.

REPORT HIGHLIGHT
The pet page in the report uses the provided marketing images as bookmark triggers to switch between a dog view and a cat view. This means that when users trigger the cat view, all data on the page (including values in slicers) contain only cat data, and the same goes for the dog data.


NOTE - This project is part of my submission for the Udacity & Microsoft PowerBI Nanodegree program

© PEACE PETER